+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { Work_Sans as FontSans } from "next/font/google";
 import "./globals.css";
 import { cn } from "@/lib/utils";
-import Header from "@/components/ui/header";
+import Header from "@/components/ui/header/header";
 import { useEffect, useState } from "react";
 import Footer from "@/components/ui/footer";
