+++ src/components/ui/headerHeader.cy.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "./header";
+import Header from "./header/header";
 
 describe("<Header />", () => {
   it("renders", () => {
